Deduplicate EXT1 implementations
