+++ esp-hal/src/rtc_cntl/sleep/esp32.rs
@@ -1,4 +1,4 @@
-use super::{Ext0WakeupSource, Ext1WakeupSource, WakeSource, WakeTriggers};
+use super::{Ext0WakeupSource, WakeSource, WakeTriggers};
 use crate::{
     gpio::{RtcFunction, RtcPin},
     peripherals::{BB, DPORT, I2S0, LPWR, NRX, RTC_IO},
@@ -106,53 +106,6 @@ impl<P: RtcPin> Drop for Ext0WakeupSource<P> {
         self.pin
             .borrow_mut()
             .rtc_set_config(true, false, RtcFunction::Rtc);
-    }
-}
-
-impl WakeSource for Ext1WakeupSource<'_, '_> {
-    fn apply(
-        &self,
-        _rtc: &Rtc<'_>,
-        triggers: &mut WakeTriggers,
-        _sleep_config: &mut RtcSleepConfig,
-    ) {
-        triggers.insert(WakeupSource::Ext1);
-
-        // set pins to RTC function
-        let mut pins = self.pins.borrow_mut();
-        let mut bits = 0u32;
-        for pin in pins.iter_mut() {
-            pin.rtc_set_config(true, true, RtcFunction::Rtc);
-            pin.rtcio_pad_hold(true);
-            bits |= 1 << pin.rtc_number();
-        }
-
-        // clear previous wakeup status
-        LPWR::regs()
-            .ext_wakeup1()
-            .modify(|_, w| w.status_clr().set_bit());
-        // set pin register field
-
-        LPWR::regs()
-            .ext_wakeup1()
-            .modify(|_, w| unsafe { w.sel().bits(bits) });
-
-        // set level register field
-        LPWR::regs()
-            .ext_wakeup_conf()
-            .modify(|_r, w| w.ext_wakeup1_lv().bit(self.level == WakeupLevel::High));
-    }
-}
-
-impl Drop for Ext1WakeupSource<'_, '_> {
-    fn drop(&mut self) {
-        // should we have saved the pin configuration first?
-        // set pin back to IO_MUX (input_enable and func have no effect when pin is sent
-        // to IO_MUX)
-        let mut pins = self.pins.borrow_mut();
-        for pin in pins.iter_mut() {
-            pin.rtc_set_config(true, false, RtcFunction::Rtc);
-        }
     }
 }
