+++ esp-hal/src/rtc_cntl/sleep/esp32s3.rs
@@ -1,11 +1,4 @@
-use super::{
-    Ext0WakeupSource,
-    Ext1WakeupSource,
-    UlpWakeupSource,
-    WakeSource,
-    WakeTriggers,
-    WakeupLevel,
-};
+use super::{Ext0WakeupSource, UlpWakeupSource, WakeSource, WakeTriggers, WakeupLevel};
 use crate::{
     gpio::{RtcFunction, RtcPin},
     peripherals::{APB_CTRL, EXTMEM, LPWR, RTC_IO, SPI0, SPI1, SYSTEM},
@@ -145,54 +138,6 @@ impl<P: RtcPin> Drop for Ext0WakeupSource<P> {
         self.pin
             .borrow_mut()
             .rtc_set_config(true, false, RtcFunction::Rtc);
-    }
-}
-
-impl WakeSource for Ext1WakeupSource<'_, '_> {
-    fn apply(
-        &self,
-        _rtc: &Rtc<'_>,
-        triggers: &mut WakeTriggers,
-        _sleep_config: &mut RtcSleepConfig,
-    ) {
-        triggers.insert(WakeupSource::Ext1);
-
-        // set pins to RTC function
-        let mut pins = self.pins.borrow_mut();
-        let mut bits = 0u32;
-        for pin in pins.iter_mut() {
-            pin.rtc_set_config(true, true, RtcFunction::Rtc);
-            pin.rtcio_pad_hold(true);
-            bits |= 1 << pin.rtc_number();
-        }
-
-        unsafe {
-            let rtc_cntl = LPWR::regs();
-            // clear previous wakeup status
-            rtc_cntl
-                .ext_wakeup1()
-                .modify(|_, w| w.ext_wakeup1_status_clr().set_bit());
-            // set pin register field
-            rtc_cntl
-                .ext_wakeup1()
-                .modify(|_, w| w.ext_wakeup1_sel().bits(bits));
-            // set level register field
-            rtc_cntl
-                .ext_wakeup_conf()
-                .modify(|_r, w| w.ext_wakeup1_lv().bit(self.level == WakeupLevel::High));
-        }
-    }
-}
-
-impl Drop for Ext1WakeupSource<'_, '_> {
-    fn drop(&mut self) {
-        // should we have saved the pin configuration first?
-        // set pin back to IO_MUX (input_enable and func have no effect when pin is sent
-        // to IO_MUX)
-        let mut pins = self.pins.borrow_mut();
-        for pin in pins.iter_mut() {
-            pin.rtc_set_config(true, false, RtcFunction::Rtc);
-        }
     }
 }
